feat(compute): Move system configs into nonvolatile storage
Instead of COPYing the configurations and setup scripts into /usr/local/bin, we
now put them in /data/system so they will persist across container relaunches.

This requires some rearchitecting of the container initialization.
Unfortunately, because /data is not mounted until the container runs, we have to
provision the setup and start scripts at runtime. To enable this, the logic that
runs on container start that used to be in setup.sh is now in a new script,
container_setup.sh, that does the bare minimum to ensure a freshly started
container (on resin) can run the server.