feat(api): Add robot reset endpoints and implementation to API (#1855)
Add endpoints at GET /settings/reset/options and PUT /settings/reset to, respectively, query what
reset options are available and reset any subset of those options to their factory configurations.
This allows us to reset only the parts of the configuration that a user might want, and avoid
resetting those configurations that are required for operation and difficult to change.